Changes to Unit infrastructure to throw exceptions when parsing units. This allows for better validation but requires more checking throughout the code.  Anywhere units are used needs to catch exceptions provide a means of operating without a unit or setting the unit to ONE.
The DataPointVO class checks during serialization and on validation.

The logic for DWRs and REST when settings units is different where the
DWR system uses the unitString memebers and converts them into Units
while the REST model converts directly.  Both check for valid units and
if they do not parse properly the corresponding unit is set to null.
The value of NULL is used to indicate an invalid unit during validation.

Also UTF-8 conversion is required for units so the Csv backend now
enforces the stream to UTF-8 when decoding.